+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [6.0.0-rc.1] - 2026-08-23
 
 ### Changed
 
@@ -794,7 +794,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
-[Unreleased]: https://https://github.com/internetguru/laravel-model-browser/compare/staging...dev
+[6.0.0-rc.1]: https://github.com/internetguru/laravel-model-browser/releases/tag/v5.3.3
 [5.3.3]: https://https://github.com/internetguru/laravel-model-browser/compare/v5.3.2...v5.3.3
 [5.3.2]: https://https://github.com/internetguru/laravel-model-browser/compare/v5.3.1...v5.3.2
 [5.3.1]: https://https://github.com/internetguru/laravel-model-browser/compare/v5.3.0...v5.3.1